backend/pyproject.toml
[tool.poetry]
name = "price-tracker"
version = "0.1.0"
description = ""
authors = ["AdamAbdulkareem <dmabdulkareem@gmail.com>"]
readme = "README.md"
packages = [{include = "price_tracker"}]

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
playwright = "^1.37.0"
playwright-stealth = "^1.0.5"
flask = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
